Update TLS package structure
