+++ pkg/tls/verifier/verifier.go
@@ -5,6 +5,18 @@ import (
 )
 
 type Verifier interface {
-	VerifyVerifiedPeerCertificates(verifiedChains [][]*x509.Certificate) error
-	VerifyRawPeerCertificates(peerCertificates []*x509.Certificate) error
+	VerifyPeerCertificate(rawCerts [][]byte, verifiedChains [][]*x509.Certificate) error
+}
+
+type Validator func(rawCerts [][]byte, verifiedChains [][]*x509.Certificate) error
+
+func NewValidator(verifiers []Verifier) Validator {
+	return func(rawCerts [][]byte, verifiedChains [][]*x509.Certificate) error {
+		for _, vm := range verifiers {
+			if err := vm.VerifyPeerCertificate(rawCerts, verifiedChains); err != nil {
+				return err
+			}
+		}
+		return nil
+	}
 }
